Agregada seguridad basica de roles a usuarios e inmuebles

+++ backend/src/modules/inmuebles/controllers/DetalleController.js
@@ -1,11 +1,15 @@
 //Manejar las solicitudes HTTP. Llama al servicio correspondiente. Este maneja las solicitudes GET
 const errorHandler = require('../../../utils/ErrorHandler');
-const detalleService = require('../services/DetalleService'); //Importar el servicio 
+const detalleService = require('../services/DetalleService');
+const CustomerService = require('../../usuarios/services/CustomerService');
+const FiltrosInmuebleService = require('../services/FiltrosInmuebleService');
 const fs = require('fs'); // Para manejar la eliminación de archivos
 const { deleteMultimediaServidor } = require('../../../middleware/uploadConfig');
+const { traerToken } = require('../../../conf/firebaseAuth');
 
 // Asignar una foto/video a un detalle
 const insertMultimedia = async (req, res) => {
+    const token = await traerToken(req);
     let rutaFoto = null;
     let nombreFoto = null;
     let tipoArchivo = null;
@@ -20,9 +24,18 @@ const insertMultimedia = async (req, res) => {
 
     try {
         const { idDetalle } = req.params;
+        /* Verificar que el customer dueño sea el mismo que inició sesion
+          Si el usuario es admin se permite el ver los datos*/
+        const idCustomer = FiltrosInmuebleService.traerCustomerInmueble(token.idUsuario, idDetalle, null);
+        if (token.tipoUsuario == "admin" || CustomerService.coincideIdUsuario(token.idUsuario, idCustomer)) {
 
-        msg = await detalleService.insertarMultimediaDetalle(idDetalle, nombreFoto, tipoArchivo);
-        return res.status(200).json({ message: msg }); // Asegúrate de enviar una respuesta aquí
+            msg = await detalleService.insertarMultimediaDetalle(idDetalle, nombreFoto, tipoArchivo);
+            return res.status(200).json({ message: msg }); // respuesta aquí
+
+        } else {
+            throw new ErrorNegocio("No tiene permisos o el id del usuario que inició sesion no coincide con el solicitado.")
+        }
+
     } catch (err) {
         // Elimina el archivo subido si hubo un error en la inserción
         if (rutaFoto) {
@@ -41,9 +54,17 @@ const insertMultimedia = async (req, res) => {
 // Eliminar un detalle
 const eliminarDetalle = async (req, res) => {
     try {
+        const token = await traerToken(req);
         const { idDetalle } = req.params;
-        msg = await detalleService.deleteDetalle(idDetalle);
-        res.status(201).json(msg); //Se retorna un mensaje
+        /* Verificar que el customer dueño sea el mismo que inició sesion
+          Si el usuario es admin se permite el ver los datos*/
+        const idCustomer = FiltrosInmuebleService.traerCustomerInmueble(token.idUsuario, idDetalle, null);
+        if (token.tipoUsuario == "admin" || CustomerService.coincideIdUsuario(token.idUsuario, idCustomer)) {
+            msg = await detalleService.deleteDetalle(idDetalle);
+            res.status(201).json(msg); //Se retorna un mensaje
+        } else {
+            throw new ErrorNegocio("No tiene permisos o el id del usuario que inició sesion no coincide con el solicitado.")
+        }
     } catch (err) {
         errorHandler.handleControllerError(res, err, "inmuebles");
     }
@@ -53,7 +74,7 @@ const eliminarDetalle = async (req, res) => {
 const eliminarMultimediaDetalle = async (req, res) => {
     try {
         let msg;
-
+        const token = await traerToken(req);
         const { idDetalle, idFoto, idVideo } = req.params;
 
         // Si idFoto existe el tipo es foto, si no es tipo video
@@ -61,10 +82,16 @@ const eliminarMultimediaDetalle = async (req, res) => {
 
         // Toma el id de foto o video
         const id = idFoto || idVideo;
+        /* Verificar que el customer dueño sea el mismo que inició sesion
+          Si el usuario es admin se permite el ver los datos*/
+        const idCustomer = FiltrosInmuebleService.traerCustomerInmueble(token.idUsuario, idDetalle, null);
+        if (token.tipoUsuario == "admin" || CustomerService.coincideIdUsuario(token.idUsuario, idCustomer)) {
+            let msg = await detalleService.deleteMultimediaBD(id, tipo, idDetalle);
 
-        msg = await detalleService.deleteMultimediaBD(id, tipo, idDetalle);
-
-        res.status(201).json(msg); //Se retorna un mensaje
+            res.status(201).json(msg); //Se retorna un mensaje
+        } else {
+            throw new ErrorNegocio("No tiene permisos o el id del usuario que inició sesion no coincide con el solicitado.")
+        }
 
     } catch (err) {
         console.log(err);
