+++ backend/src/modules/inmuebles/controllers/InmuebleController.js
@@ -10,18 +10,30 @@ const insertInmueble = async (req, res) => {
         msg = await inmuebleService.insertarInmueble(req.body);
         res.status(201).json(msg); //Se retorna un mensaje
     } catch (error) {
-       errorHandler.handleControllerError(res,error,"inmuebles");
+        errorHandler.handleControllerError(res, error, "inmuebles");
     }
 };
 
 // Insertar un inmueble
 const agregarZona = async (req, res) => {
     try {
+        const token = await traerToken(req);
         const datos = req.body; //Datos del cuerpo de la solicitud (zonas)
         datos.idInmueble = req.params.idInmueble;
-        msg = await zonasInmueblesService.agregarZona(datos);
+
+        // Traer dueño del inmueble
+        const idCustomer = filtroInmueble.traerCustomerInmueble(token.idUsuario, null, datos.idInmueble);
+
+        // Validar que sea el dueño
+        if (token.tipoUsuario == "admin" || CustomerService.coincideIdUsuario(token.idUsuario, idCustomer)) {
+            let msg = await zonasInmueblesService.agregarZona(datos);
+            res.status(201).json(msg); //Se retorna un mensaje
+        } else {
+            throw new ErrorNegocio("No tiene permisos o el id del usuario que inició sesion no coincide con el solicitado.")
+        }
+
     } catch (err) {
-        errorHandler.handleControllerError(res,err,"inmuebles");
+        errorHandler.handleControllerError(res, err, "inmuebles");
     }
 };
 
@@ -31,7 +43,7 @@ const getInmueblesPublicados = async (req, res) => {
         msg = await filtroInmueble.getPublicados(req.query);
         res.status(201).json(msg); //Se retorna un mensaje
     } catch (err) {
-        errorHandler.handleControllerError(res,err,"inmuebles");
+        errorHandler.handleControllerError(res, err, "inmuebles");
     }
 };
 
@@ -41,7 +53,7 @@ const getAllTipos = async (req, res) => {
         msg = await inmuebleService.getAllTipos();
         res.status(201).json(msg); //Se retorna un mensaje
     } catch (err) {
-        errorHandler.handleControllerError(res,err,"inmuebles");
+        errorHandler.handleControllerError(res, err, "inmuebles");
     }
 };
 
@@ -52,7 +64,7 @@ const getInmuebleByID = async (req, res) => {
         console.log("===");
         res.status(201).json(msg); //Se retorna un mensaje
     } catch (err) {
-        errorHandler.handleControllerError(res,err,"inmuebles");
+        errorHandler.handleControllerError(res, err, "inmuebles");
     }
 };
 
@@ -61,9 +73,9 @@ const getInmuebleByID = async (req, res) => {
 const getInmueblesUsuario = async (req, res) => {
     try {
         msg = await filtroInmueble.getInmueblesUsuario(req.params);
-        res.status(201).json(msg); 
+        res.status(201).json(msg);
     } catch (err) {
-        errorHandler.handleControllerError(res,err,"inmuebles");
+        errorHandler.handleControllerError(res, err, "inmuebles");
     }
 };
 
@@ -73,7 +85,7 @@ const getInmueblesCodigo = async (req, res) => {
         msg = await filtroInmueble.getInmueblesCodigo(req.params);
         res.status(201).json(msg); //Se retorna un mensaje si se encuentra un error
     } catch (err) {
-        errorHandler.handleControllerError(res,err,"inmuebles");
+        errorHandler.handleControllerError(res, err, "inmuebles");
     }
 };
 
@@ -84,7 +96,7 @@ const getInteresadosInmueble = async (req, res) => {
         interesados = await inmuebleService.traerInteresados(req.params);
         res.status(201).json(interesados); //Se retorna un mensaje si se encuentra un error
     } catch (err) {
-        errorHandler.handleControllerError(res,err,"inmuebles");
+        errorHandler.handleControllerError(res, err, "inmuebles");
     }
 };
 
@@ -92,21 +104,44 @@ const getInteresadosInmueble = async (req, res) => {
 //Editar un inmueble con detalles
 const actualizarInmuebleDetalles = async (req, res) => {
     try {
-        msg = await inmuebleService.actualizarInmuebleDetalles(req.body, req.params);
-        res.status(201).json(msg); //Se retorna un mensaje si se encuentra un error
+        const token = await traerToken(req);
+        const { idInmueble } = req.params;
+
+        // Traer dueño del inmueble
+        const idCustomer = filtroInmueble.traerCustomerInmueble(token.idUsuario, null, idInmueble);
+
+        // Validar que sea el dueño
+        if (token.tipoUsuario == "admin" || CustomerService.coincideIdUsuario(token.idUsuario, idCustomer)) {
+            let msg = await inmuebleService.actualizarInmuebleDetalles(req.body, req.params);
+            res.status(201).json(msg); //Se retorna un mensaje si se encuentra un error
+        } else {
+            throw new ErrorNegocio("No tiene permisos o el id del usuario que inició sesion no coincide con el solicitado.")
+        }
+
     } catch (err) {
-        errorHandler.handleControllerError(res,err,"inmuebles");
+        errorHandler.handleControllerError(res, err, "inmuebles");
     }
 }
 
 // eliminar inmueble
 const eliminarInmueble = async (req, res) => {
     try {
-        msg = await inmuebleService.eliminarInmueble(req.params);
-        res.status(201).json(msg); //Se retorna un mensaje si se encuentra un error
+        const token = await traerToken(req);
+        const { idInmueble } = req.params;
+
+        // Traer dueño del inmueble
+        const idCustomer = filtroInmueble.traerCustomerInmueble(token.idUsuario, null, idInmueble);
+
+        // Validar que sea el dueño
+        if (token.tipoUsuario == "admin" || CustomerService.coincideIdUsuario(token.idUsuario, idCustomer)) {
+            msg = await inmuebleService.eliminarInmueble(req.params);
+            res.status(201).json(msg); //Se retorna un mensaje si se encuentra un error
+        } else {
+            throw new ErrorNegocio("No tiene permisos o el id del usuario que inició sesion no coincide con el solicitado.")
+        }
     } catch (err) {
         console.log(err);
-        errorHandler.handleControllerError(res,err,"inmuebles");
+        errorHandler.handleControllerError(res, err, "inmuebles");
     }
 }
 module.exports = {
